[webapp] rename Telegram provider
